refactor: AgentLinkManager 拆 Reducer/Presentation（批6-5）
- AgentLinkReducer：busy/idle/attention/error 聚合、去抖、完成确认、
  代次校验、cooldown，纯状态机逐字段等价迁移
- AgentLinkPresentation：气泡/音效/动画调度，只经 PetWindow 公开 seam
- Manager 只留装配/监视器生命周期/安装卸载编排
- 5.6sol 盲审：状态机/信号语义无差异；P2（Presentation 经类属性反向
  依赖 Reducer 常量）已修为模块级 BUSY_STATES 共享常量
- 验收期间全量 3 次命中已知 flake（登记册 #2/#7/#9），基线 worktree
  同点复现证明与本批无关；focused 84 绿 ×2

结构优化批6-5（STRUCT_PLAN_FINAL.md）
